Terrain/RasterTileCache: explicit 16 bit integer for tile index
"unsigned short" is usually 16 bit, but let's use a type that is
guaranteed to be 16 bit.  Since that modifies the layout of the
cache file, the version must be incremented.

+++ src/Terrain/RasterTile.cpp
@@ -385,7 +385,7 @@ RasterTileCache::MarkerSegment(long file_offset, unsigned id)
   if (operation != NULL)
     operation->SetProgressPosition(file_offset / 65536);
 
-  int tile = MarkerSegmentInfo::NO_TILE;
+  uint16_t tile = MarkerSegmentInfo::NO_TILE;
   if (is_tile_segment(id) && !segments.empty())
     /* this segment belongs to the same tile as the preceding SOT
        segment */
